Major: created second interface to manage format_data()
