+++ sensor interface/README.md
@@ -13,9 +13,9 @@ In the examples folder is found a full implementation to a sensor: see examples/
 We have defined 3 levels in our hierarchy:
 * Sensors
 * Sensor manufacturers
-* sensor model
+* Sensor model
 
-e.g. SPEC-DGS (manufacturer) produces the CO sensor (sensor model).
+e.g. SPEC (sensor manufacturer) produces the [CO](https://www.digikey.ca/en/products/detail/spec-sensors-llc/968-034/6676880) sensor (sensor model).
 
 ### Sensors
 
@@ -29,13 +29,36 @@ SensorInterface.print_formatted_data() implements a method that should be called
 
 ### Sensor manufacturers
 
-Each Sensor manufacturer should define an abstract method, format_data(). A sensor model must implement that method to format data in an agreed-upon format. 
+Each Sensor manufacturer should define an abstract method, format_data(). A sensor model must implement that method to format data in an agreed-upon format.
+
+Note measurements should be expressed in SI units, so perform conversions at this level. See more [here](#units)
 
 ### Sensor model
 
 Here you define any model-specific behaviour. For SPEC-DGS sensors this involves implementing format_data() (to be read by SensorInterface.print_formatted_data()
 
-## Units
+## Adding a new sensor to the system
+
+We'll assume it's a new sensor from a new manufacturer.
+If the manufacturer class has already been created then only [sensor class implementation](#2-implement-the-sensor-class) is necessary.
+
+### 1. Implement a manufacturer class
+
+This class must do two things
+1. implement SensorInterface.py
+
+See spec_dgs.py for an example
+
+### 2. Implement the sensor class
+
+This class extends the manufacturer class & must implement sensor_data_formatter.py
+
+See spec_co_sensor.py for an example.
+
+
+## Related topics
+
+### Units
 
 All measurements should be expressed in [SI units](https://en.wikipedia.org/wiki/International_System_of_Units). However, dimensionless quantities, like ppm (parts per million) are not SI and have no SI definition. For consistency, we've decided to standardize to ppm.
