+++ sensor interface/sensors/__init__.py
@@ -1,3 +1,4 @@
 from .SensorInterface import SensorInterface
 from .spec_dgs import spec_dgs
 from .spec_co_sensor import spec_co_sensor
+from .sensor_data_formatter import sensor_data_formatter
