+++ sensor interface/sensors/sensor_data_formatter.py
@@ -1,0 +1,21 @@
+import abc
+
+
+class sensor_data_formatter(metaclass=abc.ABCMeta):
+    @classmethod
+    def __subclasshook__(cls, subclass):
+        return (hasattr(subclass, 'format_data') and
+                callable(subclass.format_data) or
+                NotImplemented)
+
+    @abc.abstractmethod
+    def format_data(self) -> dict:
+        f"""
+           Provides formatted sensor data for output.
+
+           Specifications:
+           - key names must be fully spelled out. e.g. 'carbon_monoxide', 'relative_humidity'
+
+           :return: {dict} 
+           """
+        raise NotImplementedError
